requirements.txt
playwright
supabase
torch
transformers
pillow
httpx
python-dotenv